+++ core/new-gui/src/app/workspace/component/property-editor/breakpoint-frame/breakpoint-frame/breakpoint-frame.component.html
@@ -1,0 +1,39 @@
+<div *ngIf="displayForm">
+  <h3 *ngIf="formTitle" class="texera-workspace-property-editor-title" mat-subheader>
+    {{formTitle}}
+  </h3>
+
+  <form *ngIf="formlyFields" [formGroup]="formlyFormGroup" class="texera-workspace-property-editor-form">
+    <formly-form (modelChange) = "onFormChanges($event)"
+                 [fields]="formlyFields"
+                 [form]="formlyFormGroup"
+                 [model]="formData"
+                 [options]="formlyOptions">
+    </formly-form>
+  </form>
+
+  <button (click)="handleAddBreakpoint()" *ngIf="currentLinkID !== undefined && !hasBreakpoint()" class="breakpointRemoveButton"
+          type="button">Add Breakpoint
+  </button>
+
+  <button (click)="handleRemoveBreakpoint()" *ngIf="currentLinkID !== undefined && hasBreakpoint()" class="breakpointRemoveButton"
+          type="button">Remove Breakpoint
+  </button>
+
+  <button (click)="allowChangeOperatorLogic()"
+          *ngIf="currentOperatorID !== undefined &&
+    (this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered) &&
+    this.interactive === false" class="breakpointRemoveButton"
+          type="button">Change Logic
+  </button>
+
+  <button (click)="confirmChangeOperatorLogic()"
+          *ngIf="currentOperatorID !== undefined &&
+    (this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered) &&
+    this.interactive === true" class="breakpointRemoveButton"
+          type="button">Confirm Change
+  </button>
+</div>
+
