+++ core/new-gui/src/app/workspace/component/property-editor/breakpoint-frame/breakpoint-frame/breakpoint-frame.component.ts
@@ -1,0 +1,298 @@
+import { Component, OnInit } from '@angular/core';
+import { cloneDeep, isEqual } from 'lodash';
+import { ExecuteWorkflowService, FORM_DEBOUNCE_TIME_MS } from '../../../../service/execute-workflow/execute-workflow.service';
+import {
+  SchemaAttribute,
+  SchemaPropagationService
+} from '../../../../service/dynamic-schema/schema-propagation/schema-propagation.service';
+import { DynamicSchemaService } from '../../../../service/dynamic-schema/dynamic-schema.service';
+import { WorkflowActionService } from '../../../../service/workflow-graph/model/workflow-action.service';
+import { FormlyJsonschema } from '@ngx-formly/core/json-schema';
+import { ExecutionState } from 'src/app/workspace/types/execute-workflow.interface';
+import { FormGroup } from '@angular/forms';
+import { Subject } from 'rxjs/Subject';
+import { Observable } from 'rxjs/Observable';
+import { FormlyFieldConfig, FormlyFormOptions } from '@ngx-formly/core';
+import { CustomJSONSchema7 } from '../../../../types/custom-json-schema.interface';
+import { setChildTypeDependency, setHideExpression } from 'src/app/common/formly/formly-utils';
+
+@Component({
+  selector: 'texera-breakpoint-frame',
+  templateUrl: './breakpoint-frame.component.html',
+  styleUrls: ['./breakpoint-frame.component.scss']
+})
+export class BreakpointFrameComponent implements OnInit {
+
+
+  // re-declare enum for angular template to access it
+  public readonly ExecutionState = ExecutionState;
+
+  // operatorID if the component is displaying operator property editor
+  public currentOperatorID: string | undefined;
+
+  // the linkID if the component is displaying breakpoint editor
+  public currentLinkID: string | undefined;
+
+  // used in HTML template to control if the form is displayed
+  public displayForm: boolean = false;
+
+  // whether the editor can be edited
+  public interactive: boolean = true;
+
+  // the source event stream of form change triggered by library at each user input
+  public sourceFormChangeEventStream = new Subject<object>();
+
+
+  public breakpointChangeStream = this.createOutputFormChangeEventStream(
+    this.sourceFormChangeEventStream, data => this.checkBreakpoint(data));
+
+  // inputs and two-way bindings to formly component
+  public formlyFormGroup: FormGroup | undefined;
+  public formData: any;
+  public formlyOptions: FormlyFormOptions | undefined;
+  public formlyFields: FormlyFieldConfig[] | undefined;
+  public formTitle: string | undefined;
+
+
+  constructor(
+    public formlyJsonschema: FormlyJsonschema,
+    public workflowActionService: WorkflowActionService,
+    public autocompleteService: DynamicSchemaService,
+    public executeWorkflowService: ExecuteWorkflowService,
+    private schemaPropagationService: SchemaPropagationService
+  ) { }
+
+  ngOnInit(): void {
+    const highlightLinks = this.workflowActionService.getJointGraphWrapper().getCurrentHighlightedLinkIDs();
+    this.showBreakpointEditor(highlightLinks[0]);
+    // when the operator's property is updated via program instead of user updating the json schema form,
+    //  this observable will be responsible in handling these events.
+    this.registerOperatorPropertyChangeHandler();
+
+    // handle the form change event on the user interface to actually set the operator property
+    this.registerOnFormChangeHandler();
+  }
+
+  /**
+   * This method handles the form change event
+   */
+  registerOnFormChangeHandler(): void {
+    // TODO: handle change
+  }
+
+  public hasBreakpoint(): boolean {
+    if (!this.currentLinkID) {
+      return false;
+    }
+    return this.workflowActionService.getTexeraGraph().getLinkBreakpoint(this.currentLinkID) !== undefined;
+  }
+
+  public handleAddBreakpoint() {
+    if (this.currentLinkID && this.workflowActionService.getTexeraGraph().hasLinkWithID(this.currentLinkID)) {
+      this.workflowActionService.setLinkBreakpoint(this.currentLinkID, this.formData);
+      if (this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
+        this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered) {
+        this.executeWorkflowService.addBreakpointRuntime(this.currentLinkID, this.formData);
+      }
+    }
+  }
+
+  /**
+   * This method handles the link breakpoint remove button click event.
+   * It will hide the property editor, clean up currentBreakpointInitialData.
+   * Then unhighlight the link and remove it from the workflow.
+   */
+  public handleRemoveBreakpoint() {
+    if (this.currentLinkID) {
+      // remove breakpoint in texera workflow first, then unhighlight it
+      this.workflowActionService.removeLinkBreakpoint(this.currentLinkID);
+      this.workflowActionService.getJointGraphWrapper().unhighlightLinks(this.currentLinkID);
+    }
+    this.clearPropertyEditor();
+  }
+
+  public allowChangeOperatorLogic() {
+    this.setInteractivity(true);
+  }
+
+  public confirmChangeOperatorLogic() {
+    this.setInteractivity(false);
+    if (this.currentOperatorID) {
+      this.executeWorkflowService.changeOperatorLogic(this.currentOperatorID);
+    }
+  }
+
+  public showBreakpointEditor(linkID: string): void {
+    if (!this.workflowActionService.getTexeraGraph().hasLinkWithID(linkID)) {
+      throw new Error(`change property editor: link does not exist`);
+    }
+    // set the operator data needed
+    this.currentLinkID = linkID;
+    const breakpointSchema = this.autocompleteService.getDynamicBreakpointSchema(linkID).jsonSchema;
+
+    this.formTitle = 'Breakpoint';
+    const breakpoint = this.workflowActionService.getTexeraGraph().getLinkBreakpoint(linkID);
+    this.formData = breakpoint !== undefined ? cloneDeep(breakpoint) : {};
+    this.setFormlyFormBinding(breakpointSchema);
+
+    // show breakpoint editor
+    this.displayForm = true;
+
+    const interactive = this.executeWorkflowService.getExecutionState().state === ExecutionState.Uninitialized ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.Completed;
+    this.setInteractivity(interactive);
+  }
+
+  /**
+   * Handles the form change event stream observable,
+   *  which corresponds to every event the json schema form library emits.
+   *
+   * Applies rules that transform the event stream to trigger reasonably and less frequently,
+   *  such as debounce time and distinct condition.
+   *
+   * Then modifies the operator property to use the new form data.
+   */
+  public createOutputFormChangeEventStream(
+    formChangeEvent: Observable<object>,
+    modelCheck: (formData: object) => boolean
+  ): Observable<object> {
+
+    return formChangeEvent
+      // set a debounce time to avoid events triggering too often
+      //  and to circumvent a bug of the library - each action triggers event twice
+      .debounceTime(FORM_DEBOUNCE_TIME_MS)
+      // .do(evt => console.log(evt))
+      // don't emit the event until the data is changed
+      .distinctUntilChanged()
+      // .do(evt => console.log(evt))
+      // don't emit the event if form data is same with current actual data
+      // also check for other unlikely circumstances (see below)
+      .filter(formData => modelCheck(formData))
+      // share() because the original observable is a hot observable
+      .share();
+
+  }
+
+  public setInteractivity(interactive: boolean) {
+    this.interactive = interactive;
+    if (this.formlyFormGroup !== undefined) {
+      if (this.interactive) {
+        this.formlyFormGroup.enable();
+      } else {
+        this.formlyFormGroup.disable();
+      }
+    }
+  }
+
+  /**
+   * Hides the form and clears all the data of the current the property editor
+   */
+  public clearPropertyEditor(): void {
+    // set displayForm to false in the very beginning
+    // hide the view first and then make everything null
+    this.displayForm = false;
+    this.currentOperatorID = undefined;
+    this.currentLinkID = undefined;
+
+    this.formlyFormGroup = undefined;
+    this.formData = undefined;
+    this.formlyFields = undefined;
+    this.formTitle = undefined;
+  }
+
+  /**
+   * Callback function provided to the Angular Json Schema Form library,
+   *  whenever the form data is changed, this function is called.
+   * It only serves as a bridge from a callback function to RxJS Observable
+   * @param event
+   */
+  onFormChanges(event: object): void {
+    this.sourceFormChangeEventStream.next(event);
+  }
+
+  private checkBreakpoint(formData: object): boolean {
+    // check if the component is displaying breakpoint
+    if (!this.currentLinkID) {
+      return false;
+    }
+    // check if the link still exists
+    const link = this.workflowActionService.getTexeraGraph().getLinkWithID(this.currentLinkID);
+    if (!link) {
+      return false;
+    }
+    // only emit change event if the form data actually changes
+    return !isEqual(formData, this.workflowActionService.getTexeraGraph().getLinkBreakpoint(link.linkID));
+  }
+
+  /**
+   * This method captures the change in operator's property via program instead of user updating the
+   *  json schema form in the user interface.
+   *
+   * For instance, when the input doesn't matching the new json schema and the UI needs to remove the
+   *  invalid fields, this form will capture those events.
+   */
+  private registerOperatorPropertyChangeHandler(): void {
+    this.workflowActionService.getTexeraGraph().getBreakpointChangeStream()
+      .filter(_ => this.currentLinkID !== undefined)
+      .filter(event => event.linkID === this.currentLinkID)
+      .filter(event => !isEqual(this.formData, this.workflowActionService.getTexeraGraph().getLinkBreakpoint(event.linkID)))
+      .subscribe(event => this.formData = cloneDeep(this.workflowActionService.getTexeraGraph().getLinkBreakpoint(event.linkID)));
+  }
+
+  private setFormlyFormBinding(schema: CustomJSONSchema7) {
+    // intercept JsonSchema -> FormlySchema process, adding custom options
+    // this requires a one-to-one mapping.
+    // for relational custom options, have to do it after FormlySchema is generated.
+    const jsonSchemaMapIntercept = (mappedField: FormlyFieldConfig, mapSource: CustomJSONSchema7): FormlyFieldConfig => {
+      // if the title is python script (for Python UDF), then make this field a custom template 'codearea'
+      if (mapSource?.description?.toLowerCase() === 'input your code here') {
+        if (mappedField.type) {
+          mappedField.type = 'codearea';
+        }
+      }
+      return mappedField;
+    };
+
+    this.formlyFormGroup = new FormGroup({});
+    this.formlyOptions = {};
+    // convert the json schema to formly config, pass a copy because formly mutates the schema object
+    const field = this.formlyJsonschema.toFieldConfig(cloneDeep(schema), { map: jsonSchemaMapIntercept });
+    field.hooks = {
+      onInit: (fieldConfig) => {
+        if (!this.interactive) {
+          fieldConfig?.form?.disable();
+        }
+      }
+    };
+
+    const schemaProperties = schema.properties;
+    const fields = field.fieldGroup;
+
+    // adding custom options, relational N-to-M mapping.
+    if (schemaProperties && fields) {
+      Object.entries(schemaProperties).forEach(([propertyName, propertyValue]) => {
+        if (typeof propertyValue === 'boolean') {
+          return;
+        }
+        if (propertyValue.toggleHidden) {
+          setHideExpression(propertyValue.toggleHidden, fields, propertyName);
+        }
+
+        if (propertyValue.dependOn) {
+          if (this.currentOperatorID) {
+            const attributes: ReadonlyArray<ReadonlyArray<SchemaAttribute> | null> | undefined =
+              this.schemaPropagationService.getOperatorInputSchema(this.currentOperatorID);
+            setChildTypeDependency(attributes, propertyValue.dependOn, fields, propertyName);
+          }
+        }
+      });
+    }
+
+    this.formlyFields = fields;
+
+  }
+
+
+}
