+++ core/new-gui/src/app/workspace/component/property-editor/formly-form-frame/formly-form-frame/formly-form-frame.component.html
@@ -1,0 +1,12 @@
+<h3 *ngIf="formTitle" class="texera-workspace-property-editor-title" mat-subheader>
+  {{formTitle}}
+</h3>
+
+<form *ngIf="formlyFields" [formGroup]="formlyFormGroup" class="texera-workspace-property-editor-form">
+  <formly-form (modelChange) = "onFormChanges($event)"
+               [fields]="formlyFields"
+               [form]="formlyFormGroup"
+               [model]="formData"
+               [options]="formlyOptions">
+  </formly-form>
+</form>
