+++ core/new-gui/src/app/workspace/component/property-editor/formly-form-frame/formly-form-frame/formly-form-frame.component.ts
@@ -1,0 +1,302 @@
+import { Component, OnInit } from '@angular/core';
+import { ExecuteWorkflowService, FORM_DEBOUNCE_TIME_MS } from '../../../../service/execute-workflow/execute-workflow.service';
+import { Subject } from 'rxjs/Subject';
+import { FormGroup } from '@angular/forms';
+import { FormlyFieldConfig, FormlyFormOptions } from '@ngx-formly/core';
+import * as Ajv from 'ajv';
+import { FormlyJsonschema } from '@ngx-formly/core/json-schema';
+import { WorkflowActionService } from '../../../../service/workflow-graph/model/workflow-action.service';
+import { cloneDeep, isEqual } from 'lodash';
+import { Observable } from 'rxjs/Observable';
+import { CustomJSONSchema7 } from '../../../../types/custom-json-schema.interface';
+import { isDefined } from '../../../../../common/util/predicate';
+import { ExecutionState } from 'src/app/workspace/types/execute-workflow.interface';
+import { DynamicSchemaService } from '../../../../service/dynamic-schema/dynamic-schema.service';
+import {
+  SchemaAttribute,
+  SchemaPropagationService
+} from '../../../../service/dynamic-schema/schema-propagation/schema-propagation.service';
+import { setChildTypeDependency, setHideExpression } from 'src/app/common/formly/formly-utils';
+
+@Component({
+  selector: 'texera-formly-form-frame',
+  templateUrl: './formly-form-frame.component.html',
+  styleUrls: ['./formly-form-frame.component.scss']
+})
+export class FormlyFormFrameComponent implements OnInit {
+
+  // re-declare enum for angular template to access it
+  public readonly ExecutionState = ExecutionState;
+
+  // operatorID if the component is displaying operator property editor
+  public currentOperatorID: string | undefined;
+
+  // whether the editor can be edited
+  public interactive: boolean = true;
+
+  // the source event stream of form change triggered by library at each user input
+  public sourceFormChangeEventStream = new Subject<object>();
+
+  // the output form change event stream after debounce time and filtering out values
+  public operatorPropertyChangeStream = this.createOutputFormChangeEventStream(
+    this.sourceFormChangeEventStream, data => this.checkOperatorProperty(data));
+
+
+  // inputs and two-way bindings to formly component
+  public formlyFormGroup: FormGroup | undefined;
+  public formData: any;
+  public formlyOptions: FormlyFormOptions | undefined;
+  public formlyFields: FormlyFieldConfig[] | undefined;
+  public formTitle: string | undefined;
+
+  // used to fill in default values in json schema to initialize new operator
+  private ajv = new Ajv({ useDefaults: true });
+
+  constructor(
+    public formlyJsonschema: FormlyJsonschema,
+    public workflowActionService: WorkflowActionService,
+    public executeWorkflowService: ExecuteWorkflowService,
+    public dynamicSchemaService: DynamicSchemaService,
+    private schemaPropagationService: SchemaPropagationService
+  ) {
+    // listen to the autocomplete event, remove invalid properties, and update the schema displayed on the form
+    this.registerOperatorSchemaChangeHandler();
+
+    // when the operator's property is updated via program instead of user updating the json schema form,
+    //  this observable will be responsible in handling these events.
+    this.registerOperatorPropertyChangeHandler();
+
+    // handle the form change event on the user interface to actually set the operator property
+    this.registerOnFormChangeHandler();
+
+    this.registerDisableEditorInteractivityHandler();
+
+  }
+
+  /**
+   * Handles the form change event stream observable,
+   *  which corresponds to every event the json schema form library emits.
+   *
+   * Applies rules that transform the event stream to trigger reasonably and less frequently,
+   *  such as debounce time and distinct condition.
+   *
+   * Then modifies the operator property to use the new form data.
+   */
+  public createOutputFormChangeEventStream(
+    formChangeEvent: Observable<object>,
+    modelCheck: (formData: object) => boolean
+  ): Observable<object> {
+
+    return formChangeEvent
+      // set a debounce time to avoid events triggering too often
+      //  and to circumvent a bug of the library - each action triggers event twice
+      .debounceTime(FORM_DEBOUNCE_TIME_MS)
+      // .do(evt => console.log(evt))
+      // don't emit the event until the data is changed
+      .distinctUntilChanged()
+      // .do(evt => console.log(evt))
+      // don't emit the event if form data is same with current actual data
+      // also check for other unlikely circumstances (see below)
+      .filter(formData => modelCheck(formData))
+      // share() because the original observable is a hot observable
+      .share();
+
+  }
+
+  ngOnInit(): void {
+    const highlightedOperators = this.workflowActionService.getJointGraphWrapper().getCurrentHighlightedOperatorIDs();
+    if (highlightedOperators.length !== 1) {
+      return;
+    }
+    const currentOperatorId = highlightedOperators[0];
+    this.showOperatorPropertyEditor(currentOperatorId);
+  }
+
+  /**
+   * Callback function provided to the Angular Json Schema Form library,
+   *  whenever the form data is changed, this function is called.
+   * It only serves as a bridge from a callback function to RxJS Observable
+   * @param event
+   */
+  onFormChanges(event: object): void {
+    this.sourceFormChangeEventStream.next(event);
+  }
+
+  /**
+   * Changes the property editor to use the new operator data.
+   * Sets all the data needed by the json schema form and displays the form.
+   * @param operatorId
+   */
+  showOperatorPropertyEditor(operatorId: string): void {
+    const operator = this.workflowActionService.getTexeraGraph().getOperator(operatorId);
+    // set the operator data needed
+    this.currentOperatorID = operatorId;
+    const currentOperatorSchema = this.dynamicSchemaService.getDynamicSchema(this.currentOperatorID);
+    this.setFormlyFormBinding(currentOperatorSchema.jsonSchema);
+    this.formTitle = currentOperatorSchema.additionalMetadata.userFriendlyName;
+
+    /**
+     * Important: make a deep copy of the initial property data object.
+     * Prevent the form directly changes the value in the texera graph without going through workflow action service.
+     */
+    this.formData = cloneDeep(operator.operatorProperties);
+
+    // use ajv to initialize the default value to data according to schema, see https://ajv.js.org/#assigning-defaults
+    // WorkflowUtil service also makes sure that the default values are filled in when operator is added from the UI
+    // However, we perform an addition check for the following reasons:
+    // 1. the operator might be added not directly from the UI, which violates the precondition
+    // 2. the schema might change, which specifies a new default value
+    // 3. formly doesn't emit change event when it fills in default value, causing an inconsistency between component and service
+
+    this.ajv.validate(currentOperatorSchema, this.formData);
+
+    // manually trigger a form change event because default value might be filled in
+    this.onFormChanges(this.formData);
+
+    const interactive = this.executeWorkflowService.getExecutionState().state === ExecutionState.Uninitialized ||
+      this.executeWorkflowService.getExecutionState().state === ExecutionState.Completed;
+    this.setInteractivity(interactive);
+  }
+
+  setInteractivity(interactive: boolean) {
+    this.interactive = interactive;
+    if (this.formlyFormGroup !== undefined) {
+      this.interactive ? this.formlyFormGroup.enable() : this.formlyFormGroup.disable();
+    }
+  }
+
+  checkOperatorProperty(formData: object): boolean {
+    // check if the component is displaying operator property
+    if (this.currentOperatorID === undefined) {
+      return false;
+    }
+    // check if the operator still exists, it might be deleted during debounce time
+    const operator = this.workflowActionService.getTexeraGraph().getOperator(this.currentOperatorID);
+    if (!operator) {
+      return false;
+    }
+    // only emit change event if the form data actually changes
+    return !isEqual(formData, operator.operatorProperties);
+  }
+
+  /**
+   * This method handles the schema change event from autocomplete. It will get the new schema
+   *  propagated from autocomplete and check if the operators' properties that users input
+   *  previously are still valid. If invalid, it will remove these fields and triggered an event so
+   *  that the user interface will be updated through registerOperatorPropertyChangeHandler() method.
+   *
+   * If the operator that experiences schema changed is the same as the operator that is currently
+   *  displaying on the property panel, this handler will update the current operator schema
+   *  to the new schema.
+   */
+  registerOperatorSchemaChangeHandler(): void {
+    this.dynamicSchemaService.getOperatorDynamicSchemaChangedStream().subscribe(
+      event => {
+        if (event.operatorID === this.currentOperatorID) {
+          const currentOperatorSchema = this.dynamicSchemaService.getDynamicSchema(this.currentOperatorID);
+          const operator = this.workflowActionService.getTexeraGraph().getOperator(event.operatorID);
+          if (!operator) {
+            throw new Error(`operator ${event.operatorID} does not exist`);
+          }
+          this.setFormlyFormBinding(currentOperatorSchema.jsonSchema);
+        }
+      }
+    );
+  }
+
+  /**
+   * This method captures the change in operator's property via program instead of user updating the
+   *  json schema form in the user interface.
+   *
+   * For instance, when the input doesn't matching the new json schema and the UI needs to remove the
+   *  invalid fields, this form will capture those events.
+   */
+  registerOperatorPropertyChangeHandler(): void {
+    this.workflowActionService.getTexeraGraph().getOperatorPropertyChangeStream()
+      .filter(_ => this.currentOperatorID !== undefined)
+      .filter(operatorChanged => operatorChanged.operator.operatorID === this.currentOperatorID)
+      .filter(operatorChanged => !isEqual(this.formData, operatorChanged.operator.operatorProperties))
+      .subscribe(operatorChanged => this.formData = cloneDeep(operatorChanged.operator.operatorProperties));
+  }
+
+  /**
+   * This method handles the form change event and set the operator property
+   *  in the texera graph.
+   */
+  registerOnFormChangeHandler(): void {
+    this.operatorPropertyChangeStream.subscribe(formData => {
+      // set the operator property to be the new form data
+      if (this.currentOperatorID) {
+        this.workflowActionService.setOperatorProperty(this.currentOperatorID, cloneDeep(formData));
+      }
+    });
+  }
+
+  registerDisableEditorInteractivityHandler(): void {
+    this.executeWorkflowService.getExecutionStateStream().subscribe(event => {
+      if (this.currentOperatorID) {
+        if (event.current.state === ExecutionState.Completed || event.current.state === ExecutionState.Failed) {
+          this.setInteractivity(true);
+        } else {
+          this.setInteractivity(false);
+        }
+      }
+    });
+  }
+
+  setFormlyFormBinding(schema: CustomJSONSchema7) {
+    // intercept JsonSchema -> FormlySchema process, adding custom options
+    // this requires a one-to-one mapping.
+    // for relational custom options, have to do it after FormlySchema is generated.
+    const jsonSchemaMapIntercept = (mappedField: FormlyFieldConfig, mapSource: CustomJSONSchema7): FormlyFieldConfig => {
+      // if the title is python script (for Python UDF), then make this field a custom template 'codearea'
+      if (mapSource?.description?.toLowerCase() === 'input your code here') {
+        if (mappedField.type) {
+          mappedField.type = 'codearea';
+        }
+      }
+      return mappedField;
+    };
+
+    this.formlyFormGroup = new FormGroup({});
+    this.formlyOptions = {};
+    // convert the json schema to formly config, pass a copy because formly mutates the schema object
+    const field = this.formlyJsonschema.toFieldConfig(cloneDeep(schema), { map: jsonSchemaMapIntercept });
+    field.hooks = {
+      onInit: (fieldConfig) => {
+        if (!this.interactive) {
+          fieldConfig?.form?.disable();
+        }
+      }
+    };
+
+    const schemaProperties = schema.properties;
+    const fields = field.fieldGroup;
+
+    // adding custom options, relational N-to-M mapping.
+    if (schemaProperties && fields) {
+      Object.entries(schemaProperties).forEach(([propertyName, propertyValue]) => {
+        if (typeof propertyValue === 'boolean') {
+          return;
+        }
+        if (propertyValue.toggleHidden) {
+          setHideExpression(propertyValue.toggleHidden, fields, propertyName);
+        }
+
+        if (propertyValue.dependOn) {
+          if (isDefined(this.currentOperatorID)) {
+            const attributes: ReadonlyArray<ReadonlyArray<SchemaAttribute> | null> | undefined =
+              this.schemaPropagationService.getOperatorInputSchema(this.currentOperatorID);
+            setChildTypeDependency(attributes, propertyValue.dependOn, fields, propertyName);
+          }
+
+        }
+      });
+    }
+
+    this.formlyFields = fields;
+
+  }
+
+
+}
