+++ core/new-gui/src/app/workspace/component/property-editor/property-editor.component.html
@@ -1,39 +1,5 @@
 <div class="texera-workspace-property-editor-body texera-workflow-component-body">
-  <div *ngIf="displayForm">
-    <h3 *ngIf="formTitle" class="texera-workspace-property-editor-title" mat-subheader>
-      {{formTitle}}
-    </h3>
+  <ndc-dynamic [ndcDynamicComponent]="frameComponent"></ndc-dynamic>
+  <texera-type-casting-display [operatorID]='currentOperatorID'></texera-type-casting-display>
 
-    <form class="texera-workspace-property-editor-form" [formGroup]="formlyFormGroup" *ngIf="formlyFields">
-      <formly-form [model]="formData"
-      [fields]="formlyFields"
-      [options]="formlyOptions"
-      [form]="formlyFormGroup"
-      (modelChange) = "onFormChanges($event)">
-      </formly-form>
-    </form>
-
-    <texera-type-casting-display [operatorID]='currentOperatorID' ></texera-type-casting-display>
-
-
-    <button *ngIf="currentLinkID !== undefined && !hasBreakpoint()" type="button" class="breakpointRemoveButton"
-    (click)="handleAddBreakpoint()">Add Breakpoint</button>
-
-    <button *ngIf="currentLinkID !== undefined && hasBreakpoint()" type="button" class="breakpointRemoveButton"
-    (click)="handleRemoveBreakpoint()">Remove Breakpoint</button>
-
-    <button *ngIf="currentOperatorID !== undefined &&
-    (this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
-      this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered) &&
-    this.interactive === false"
-    type="button" class="breakpointRemoveButton"
-    (click)="allowChangeOperatorLogic()">Change Logic</button>
-
-    <button *ngIf="currentOperatorID !== undefined &&
-    (this.executeWorkflowService.getExecutionState().state === ExecutionState.Paused ||
-      this.executeWorkflowService.getExecutionState().state === ExecutionState.BreakpointTriggered) &&
-    this.interactive === true"
-    type="button" class="breakpointRemoveButton"
-    (click)="confirmChangeOperatorLogic()">Confirm Change</button>
-  </div>
 </div>
