+++ core/new-gui/src/app/workspace/service/execute-workflow/execute-workflow.service.ts
@@ -22,6 +22,7 @@ import { isEqual } from 'lodash';
 import { PAGINATION_INFO_STORAGE_KEY, ResultPaginationInfo } from '../../types/result-table.interface';
 import { sessionGetObject, sessionSetObject } from '../../../common/util/storage';
 
+// TODO: change this declaration
 export const FORM_DEBOUNCE_TIME_MS = 150;
 
 export const EXECUTE_WORKFLOW_ENDPOINT = 'queryplan/execute';
@@ -342,7 +343,7 @@ export class ExecuteWorkflowService {
    *  where each operator has an operatorID and operatorType along with
    *  the properties of the operator.
    *
-   * All the links in the workflowGraph will be tranformed to LogicalLink objects,
+   * All the links in the workflowGraph will be transformed to LogicalLink objects,
    *  where each link will store its source id as its origin and target id as its destination.
    *
    * @param workflowGraph
@@ -390,5 +391,6 @@ export class ExecuteWorkflowService {
     }
     return {operatorID, breakpoint};
   }
+
 
 }
